requirements.txt
protobuf
pytea2
pillow
